move rnafusion fixtures
