+++ tests/meta/workflow/test_rnafusion.py
@@ -7,19 +7,6 @@ from cg.models.cg_config import CGConfig
 from cg.models.rnafusion.analysis import RnafusionAnalysis
 
 from cg.meta.workflow.rnafusion import RnafusionAnalysisAPI
-from tests.cli.workflow.rnafusion.conftest import (
-    fixture_rnafusion_context,
-    fixture_rnafusion_housekeeper,
-    fixture_rnafusion_housekeeper_dir,
-    fixture_rnafusion_multiqc_json_metrics,
-    fixture_rnafusion_mock_fastq_files,
-    fixture_rnafusion_fastq_file_l_1_r_1,
-    fixture_rnafusion_fastq_file_l_1_r_2,
-    fixture_rnafusion_sample_id,
-    fixture_rnafusion_case_id,
-    fixture_no_sample_case_id,
-    mock_analysis_finish,
-)
 
 
 def test_parse_analysis(
